requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
scikit-learn>=1.5.0
joblib>=1.3.2
gunicorn>=21.2.0
python-dotenv>=1.0.0
requests>=2.31.0